#143 Remove UserService's interface
